requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.339
elasticsearch==8.17.1
python-dotenv==1.0.0
pandas==2.1.3
pydantic==2.4.2
python-multipart==0.0.6
jinja2==3.1.2
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
alibabacloud-dashscope>=1.0.0 